+++ libqpdf/QPDFWriter.cc
@@ -230,7 +230,7 @@ QPDFWriter::setEncryptionParameters(
     bits_to_clear.insert(1);
     bits_to_clear.insert(2);
 
-    unsigned long P = 0;
+    int P = 0;
     // Create the complement of P, then invert.
     for (std::set<int>::iterator iter = bits_to_clear.begin();
 	 iter != bits_to_clear.end(); ++iter)
